+++ tests/test_toolbar.py
@@ -7,6 +7,7 @@ from django.core.urlresolvers import reverse
 from django.test.utils import override_settings
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
+from djangocms_helper.utils import CMS_34
 
 from djangocms_page_sitemap.cms_toolbars import PAGE_SITEMAP_MENU_TITLE
 from djangocms_page_sitemap.models import PageSitemapProperties
@@ -37,7 +38,11 @@ class ToolbarTest(BaseTest):
         toolbar = CMSToolbar(request)
         toolbar.get_left_items()
         page_menu = toolbar.find_items(Menu, name='Page')
-        self.assertEqual(page_menu, [])
+
+        if CMS_34:
+            self.assertEqual(len(page_menu), 0)
+        else:
+            self.assertEqual(len(page_menu), 1)
 
     def test_perm(self):
         """
@@ -73,7 +78,10 @@ class ToolbarTest(BaseTest):
         toolbar = CMSToolbar(request)
         toolbar.get_left_items()
         page_menu = toolbar.find_items(Menu, name='Page')
-        self.assertEqual(page_menu, [])
+        if CMS_34:
+            self.assertEqual(len(page_menu), 0)
+        else:
+            self.assertEqual(len(page_menu), 1)
 
     def test_toolbar(self):
         """